SmartQueue/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
attrs==25.4.0
click==8.3.1
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
fastapi==0.121.0
h11==0.16.0
httptools==0.7.1
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mysqlclient==2.2.7
packaging==25.0
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sniffio==1.3.1
sqlparse==0.5.4
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
